requirements.txt
pdfrw2==0.5.0
Pillow==11.1.0
PyPDF4==1.27.0
PyYAML==6.0.2
reportlab==4.3.1
click==8.1.8
